backend/requirements.txt
# Hermes Saze Sabz - Django Backend

Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
whitenoise>=6.6.0
drf-yasg>=1.21.7
django-filter>=23.5
reportlab>=4.0.8
celery>=5.3.0
redis>=5.0.0
